Export parsers to external lib.
I created the form parser and put all the html parsers in their own lib.
