--- parse_test/form.html
@@ -1,46 +1,0 @@
-<html><head><title>Form Test</title></head><body>
-<p>Form 1:<br>
-<form action='#test' method='get'>
-  <input type='text' name='textbox' value='demo' /><br>
-  <input type='password' name='password' value='password' /><br>
-  <input type='radio' name='radio' value='value1'>value1<br>
-  <input type='radio' name='radio' value='value2' checked>value2<br>
-  <input type='radio' name='radio2' value='value3'>value3<br>
-  <input type='radio' name='radio2' value='value4' checked>value4<br>
-  <input type='checkbox' name='check' value='check1' checked>check1<br>
-  <input type='checkbox' name='check' value='check2' checked>check2<br>
-  <input type='checkbox' name='check' value='check3'>check3<br>
-  <input type='checkbox' name='check2' value='test1'>test1<br>
-  <input type='checkbox' name='check2' value='test2'>test2<br>
-  <input type="email" name="email" value='email@address.com'><br>
-  <input type="month" name="bdaymonth"><br>
-  <textarea name='message'>
-    message goes here
-  </textarea><br>
-  <textarea name='message2'>
-    here's a message
-  </textarea><br>
-  <input type='submit' />
-</form></p>
-<p>Form 2:<br>
-<form action='#test2' method='put'>
-  <input type='text' name='textbox2' value='demo2' /><br>
-  <select name='colors'>
-    <option value='red'>red</option>
-    <option value='green'>green</option>
-    <option value='blue'>blue</option>
-  </select><br>
-  <select name='names'>
-    <option value='hank'>red</option>
-    <option value='erin'>green</option>
-    <option value='jim'>blue</option>
-  </select><br>
-  <textarea name='message3'>
-    another one
-  </textarea><br>
-  <input type='date' name='birthday'><br>
-  <input type="datetime-local" name="bdaytime"><br>
-  <input type='submit' />
-</form>
-</p>
-</body></html>
